requirements.txt
requests
python-dotenv
PyYAML
pandas
pyspark
pytest